[+] Complete 04-chaining-selectors and saw Solution to check
I checked the solution because in the README.md it said I would have to
modify the html file. But I visually checked the html solution and it looks
like there was no need to change what was provided. The only thing that needed
to be changed was style.css to achieve the desired outcome.