[+] Complete 02-margin-and-padding-2
Found it difficult to recreate the outer white box.
Attempted to do so by creating a div css definition, but couldn't get it to work.
Used a body definition and it worked out. I made most classes inherit padding from
the body definition since it was tedious to do them all seperately. The exception
is the button def, which I looked at the solution to center align. 'text-align:
center' and 'align: center'. 'margin: 0 auto' was from solution and
ultimately used in source code.